[Refactor#85] HomeTabUiState를 model 패키지로 이동
